chore(#ip-145): separated handler from index
